FileDoc/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
httpx==0.26.0
anthropic==0.8.1
python-dotenv==1.0.1
gradio==4.44.1
jinja2==3.1.3
pydantic==2.6.1
tiktoken==0.6.0  # Add this
openai==1.12.0   # Add this since you're using AsyncOpenAI
streamlit==1.32.2
